Disable form edits

+++ src/components/DeleteButton.tsx
@@ -3,11 +3,13 @@ import { deleteEntryById } from "@/services/deleteById";
 import { Dispatch, SetStateAction } from "react";
 
 interface deleteButtonProps {
-    id: websiteContents["id"]    
+    id: websiteContents["id"],
+    isAdmin: boolean    
 }
 
-export const DeleteButton = ({ id }: deleteButtonProps) => {
+export const DeleteButton = ({ id, isAdmin }: deleteButtonProps) => {
   const handleClick = async (): Promise<void|Error> => {
+    if (!isAdmin) return
     try {
         deleteEntryById(id)
     } catch (error) {
@@ -15,5 +17,5 @@ export const DeleteButton = ({ id }: deleteButtonProps) => {
     }
   };
 
-  return <button onClick={handleClick}>Delete Contents</button>;
+  return <button onClick={handleClick} disabled={!isAdmin}>Delete Contents</button>;
 };
